VRUI에서 비즈니스 로직 RentalShop 으로 이동
